Shadow map sector mesh updates are here! A separate mesh is now used for shadow mapping. It is in a fast static buffer, has no face info bits, and includes edge geometry. All sector vertices are also now sent to the GPU as ivec3s, instead of vec3s (less conversion for the GPU that way). Change all places where I used the name 'face_meshes' to say 'face_mesh', as that's more accurate. Make an 'excluded' list correct. Some minor refactoring here and there too. Remove a big old comment about sector shadowing in sector.c, as pretty much everything there has been implemented.

+++ assets/shaders/sector.vert
@@ -2,7 +2,7 @@
 
 #include "common/world_shading.vert"
 
-layout(location = 0) in vec3 vertex_pos_world_space;
+layout(location = 0) in ivec3 vertex_pos_world_space;
 layout(location = 1) in uint face_info_bits;
 
 uniform bool branch;
