+++ assets/shaders/shadow/sector_depth.vert
@@ -2,7 +2,7 @@
 
 #include "../common/quad_utils.vert"
 
-layout(location = 0) in vec3 vertex_pos_world_space;
+layout(location = 0) in ivec3 vertex_pos_world_space;
 
 void main(void) {
 	gl_Position = vec4(vertex_pos_world_space, 1.0f);
